Fix: Fixed bug with Kache expired time for videos

+++ ocmsdk/src/main/java/com/gigigo/orchextra/core/data/api/dto/video/ApiVideoData.kt
@@ -4,7 +4,7 @@ import com.mskn73.kache.Kacheable
 import com.mskn73.kache.annotations.KacheLife
 import gigigo.com.vimeolibs.VimeoInfo
 
-@KacheLife(expiresTime = 1000 * 60 * 60 * 24) // 1 day
+@KacheLife(expiresTime = (1000 * 60 * 60 * 24).toLong()) // 1 day
 class ApiVideoData(val element: VimeoInfo) : Kacheable {
 
   override val key: String
